SPRT for last commit
The last commit changed the exploration factor parameter, which I forgot to mention in the commit message. The exploration factor was originally 16, but is now 4. Below is the SPRT of this change.

SPRT with elo0=0 elo1=10 alpha=0.05 beta=0.05 tc=3+0.03:
Score of auroradev vs aurorabase: 198 - 108 - 59  [0.623] 365
...      auroradev playing White: 103 - 47 - 33  [0.653] 183
...      auroradev playing Black: 95 - 61 - 26  [0.593] 182
...      White vs Black: 164 - 142 - 59  [0.530] 365
Elo difference: 87.5 +/- 33.5, LOS: 100.0 %, DrawRatio: 16.2 %
SPRT: llr 2.95 (100.2%), lbound -2.94, ubound 2.94 - H1 was accepted