Upload tune3-eval1.auroraeval
Upload tune3-eval1.auroraeval. tune3 is the name for the tuning run which is the same as tune2, but started with all parameters at 0 rather than with tune1 parameters, and has a higher learning rate of 0.01 compared to 0.0025.

This evalfile has better training error than the current tuned HCE evaluation in the main branch.
error: 0.00824585